Move generate_*_class functions from persistence/sqlalchemy/models/ to layer.py
